Added extra login and user code.
Marked locations are not showing on the map, but all data is received correctly.

Login and create account function correctly.
Points needed to be transferred to create account so that it can be sent to the database.
Login can received saved points from the database.
Someone needs to handle redeeming coupons in the application and update Tang's part to show the actual coupons and information.
I will work the script for redeeming coupons.

Signed-off-by: Jacky <jli.86@hotmail.com>